refator: use sqlx query macros

+++ rest-server/src/json/question.rs
@@ -7,9 +7,9 @@ pub struct QuestionFilter {
     #[garde(length(min = 1, max = 10))]
     pub tags: Option<Vec<i16>>,
     #[garde(range(min = 2015, max = 2024))]
-    pub year: Option<i16>,
+    pub year: Option<i32>,
     #[garde(skip)]
-    pub origin: Option<i16>,
+    pub origin: Option<i32>,
     #[garde(skip)]
     pub difficulty: Option<DifficultyLevel>,
     #[serde(flatten)]
@@ -20,7 +20,7 @@ pub struct QuestionFilter {
 #[derive(Deserialize, Serialize, Validate, Debug)]
 pub struct Pagination {
     #[garde(range(min=1, max=50))]
-    pub page: i16,
+    pub page: i64,
     #[garde(range(min=5, max=50))]
-    pub per_page: i16,
+    pub per_page: i64,
 }
